the report
[Verse 1]
Same old story, time again
Got so close but then you lost it
Should've listened to your friends
'Bout his girlfriend back in Boston
You sent him pictures and playlists and phone sex
He disappeared from the second that you said
"Let's get coffee, let's meet up"
I'm so sick of online love
And I don't understand
Why can't any man

[Pre-Chorus]
Hit it like, get it hot
Make a bitch, it's a fem
Hit it like, get it hot
Make a bitch, it's a—
Um, can you play a song with a fucking beat?

[Chorus]
(Hit it like) Hit it like rom-pom-pom-pom
(Get it hot) Get it hot like Papa John
(Make a bitch) Make a bitch go on and on
(It's a fem) It's a femininomenon
(Hit it like) Hit it like rom-pom-pom-pom
(Get it hot) Get it hot like Papa John
(Make a bitch) Make a bitch go on and on
(It's a fem) It's a femininomenon

[Verse 2]
So let's say it's working out
You pretend to love his mother
Lying to your friends about
How he's such a goddamn good lover
Stuck in the suburbs, you're folding his laundry
Got what you wanted so stop feeling sorry
Crying at the nail salon
I'm so sick of online love
And I don't understand
Why can't any man

[Pre-Chorus]
Hit it like, get it hot
Make a bitch, it's a fem
Hit it like, get it hot
Make a bitch, it's a—
Dude, can you play a song with a fucking beat?

[Chorus]
(Hit it like) Hit it like rom-pom-pom-pom
(Get it hot) Get it hot like Papa John
(Make a bitch) Make a bitch go on and on
(It's a fem) It's a femininomenon
(Hit it like) Hit it like rom-pom-pom-pom
(Get it hot) Get it hot like Papa John
(Make a bitch) Make a bitch go on and on
(It's a fem) It's a femininomenon
[Bridge]
Ladies, you know what I mean
And you know what you need
And so does he
But does it happen? (No)
But does it happen? (No)
Well, what we really need is a femininomenon
(A what?) A femininomenon

[Pre-Chorus]
(Hit it like) Hit it like rom-pom-pom-pom
(Get it hot) Get it hot like Papa John
(Make a bitch) Make a bitch go on and on
(It's a fem) It's a femininomenon
(Hit it like) Hit it like rom-pom-pom-pom
(Get it hot) Get it hot like Papa John
(Make a bitch) Make a bitch go on and on
Did you hear me? Play the fucking beat

[Chorus]
Hit it like rom-pom-pom-pom
(Get it hot) Get it hot like Papa John
(Make a bitch) Make a bitch go on and on
(It's a fem) It's a femininomenon
(Hit it like) Hit it like rom-pom-pom-pom
(Get it hot) Get it hot like Papa John
(Make a bitch) Make a bitch go on and on
(It's a fem) It's a femininomenon
[Outro]
Hit it like, get it hot
Make a bitch, it's a fem (It's a femininomenon)
Hit it like, get it hot
Make a bitch, it's a fem, fem, fem, fem